Rename Google Maps API file
